chunk: Add as_bytes() routine
Also make public some functions which are used by the Png struct
implementation.

While we are here, make the data buffer calculation account for the fact
that the input buffer may have > 1 chunk, so we should not count from
the end, but instead count from the start of the current chunk.

Same goes while calculating the CRC.